refactor: use list/command finders/scanners
Replaces 5 custom finders/scanners with default config that just calls
the list and command scanners. Primary motivation here is to dogfood the
extension mechanisms, but the obvious secondary benefit is that it DRYs
up a lot of duplication.

This breaks the scanner benchmarks so I've commented them out. I will
fix them in the next commit (or soon after).